# Conflicts:
#	src/GUI/FormThemSua/ThemSuaSanPham.java

+++ src/BUS/SanPhamService.java
@@ -100,7 +100,7 @@ public class SanPhamService {
         for (int i = result.size() - 1; i >= 0; i--) {
             SanPham sp = result.get(i);
             int soluong = sp.getSoLuong();
-            float gia = sp.getDonGia();
+            double gia = sp.getDonGia();
             Boolean soLuongKhongThoa = (soluong1 != -1 && soluong < soluong1) || (soluong2 != -1 && soluong > soluong2);
             Boolean giaKhongThoa = (gia1 != -1 && gia < gia1) || (gia2 != -1 && gia > gia2);
